resize graph with window
