+++ src/libs/d3/useChartDimensions.ts
@@ -1,29 +1,24 @@
-import { ResizeObserver } from '@juggle/resize-observer';
 import { D3ChartSettings, D3ChartSettingsProps } from './types';
 import { RefObject, useEffect, useRef, useState } from 'react';
 
 export const useChartDimensions = (
   settings: D3ChartSettingsProps
-): [RefObject<HTMLDivElement>, D3ChartSettings] => {
-  const ref = useRef<HTMLDivElement>(null);
+): [RefObject<SVGSVGElement>, D3ChartSettings] => {
+  const ref = useRef<SVGSVGElement>(null);
   const dimensions = combineChartDimensions(settings);
   const [width, setWidth] = useState(0);
   const [height, setHeight] = useState(0);
 
   useEffect(() => {
-    if (dimensions.width && dimensions.height) {
-      return;
-    }
+    if (dimensions.width && dimensions.height) return;
     const element = ref.current;
     if (!element) return;
-
     const resizeObserver = new ResizeObserver((entries) => {
-      if (!Array.isArray(entries)) return;
       if (!entries.length) return;
-      const entry = entries[0];
-      if (width !== entry.contentRect.width) setWidth(entry.contentRect.width);
-      if (height !== entry.contentRect.height)
-        setHeight(entry.contentRect.height);
+      const rect = entries[0].target?.getBoundingClientRect();
+      if (!rect) return;
+      if (width !== rect.width) setWidth(rect.width);
+      if (height !== rect.height) setHeight(rect.height);
     });
     resizeObserver.observe(element);
     return () => resizeObserver.unobserve(element);
